scripts/migrations/2020-10-miro-migration/requirements.txt
boto3
click
elasticsearch
tqdm
attrs
wellcome_storage_service